% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsimu.R
\name{windsimu}
\alias{windsimu}
\title{Simulate Datasets with User-Defined Patterns of Within-Individual Variation}
\usage{
windsimu(how_many_L2_per_L3, how_many_L1_per_L2, time_var_range,
  var_names = c("cons", "age"), x_matrix_A_cons_time = c(TRUE, TRUE),
  z_matrix_A_cons_time = c(TRUE, TRUE), x_matrix_B_cons_time = c(TRUE,
  TRUE), z_matrix_B_cons_time = c(TRUE, FALSE), beta, alpha, sigma2u,
  sigma2e = NULL, log_sigma2e = TRUE, linear_sigma2e_attempts = 10,
  random_seed)
}
\arguments{
\item{how_many_L2_per_L3}{Sample size at level 2 (there isn't an L3, but future compatible in case there ever is!)}

\item{how_many_L1_per_L2}{Sample size at level 1, per level 2 unit}

\item{time_var_range}{Vector with min and max for time variable (e.g. \code{c(12, 23)} to range from 12 to 23).}

\item{var_names}{Character vector containing names for intercept and time variable (e.g. \code{c("cons", "age")}).}

\item{x_matrix_A_cons_time}{Logical vector indicating whether intercept (cons) and time variables to be included in X design matrix A
(i.e. fixed part of model for mean of y); e.g. \code{c(TRUE, FALSE)} indicates intercept, but not time, to be included.}

\item{z_matrix_A_cons_time}{As for \code{x_matrix_A_cons_time}, but Z matrix.}

\item{x_matrix_B_cons_time}{As for \code{x_matrix_A_cons_time}, but X matrix for level 1 variance function.}

\item{z_matrix_B_cons_time}{As for \code{x_matrix_A_cons_time}, but Z matrix for level 1 variance function.}

\item{beta}{Vector containing mean for each beta (i.e. i.e. coefficients of fixed effects for model for mean of y).}

\item{alpha}{Vector containing mean for each alpha (i.e. i.e. coefficients of fixed effects for level 1 variance function).}

\item{sigma2u}{Level 2 covariance matrix; e.g. \code{matrix(c(93, 12, 2.5, 12, 3.8, 0.8, 2.5, 0.8, 0.4), nrow = 3, ncol = 3, byrow = TRUE)}.}

\item{sigma2e}{Vector (one element); only specify if 1-level model; defaults to \code{NULL}.}

\item{log_sigma2e}{Logical vector indicating whether log link used for level 1 variance function (\code{TRUE}) or not (\code{FALSE}); defaults to \code{TRUE}.}

\item{linear_sigma2e_attempts}{Maximum number of attempts made to generate non-negative sigma2e (only applicable when \code{log_sigma2e = FALSE}); defaults to \code{10}.}

\item{random_seed}{Random seed adopted.}
}
\description{
Function to create simulated dataset given user-defined parameters and data
structure for growth curve model with optional random effect for L2 variance
}
\examples{
\dontrun{
library(windsimu)

simu_u2j_covary <- within_ind_var_simu_A(
how_many_L2_per_L3 = 1000,
how_many_L1_per_L2 = 9,
x1_range = c(-1, 1),
var_names = c("cons", "age"),
x_matrix_A_x0_x1 = c(TRUE, TRUE),
z_matrix_A_x0_x1 = c(TRUE, TRUE),
x_matrix_B_x0_x1 = c(TRUE, TRUE),
z_matrix_B_x0_x1 = c(TRUE, FALSE),
beta = c(150, 6.5),
alpha = c(-.95, 0.48),
sigma2u =
  matrix(
    c(93, 12, 2.5,
      12, 3.8, 0.8,
      2.5, 0.8, 0.4),
    nrow = 3,
    ncol = 3,
    byrow = TRUE
  ),
log_sigma2eij = TRUE,
linear_sigma2eij_attempts = 10,
random_seed = 1
)

simu_u2j_no_covary <- windsimu(
how_many_L2_per_L3 = 1000,
how_many_L1_per_L2 = 9,
time_var_range = c(-1, 1),
var_names = c("cons", "age"),
x_matrix_A_cons_time = c(TRUE, TRUE),
z_matrix_A_cons_time = c(TRUE, TRUE),
x_matrix_B_cons_time = c(TRUE, TRUE),
z_matrix_B_cons_time = c(TRUE, FALSE),
beta = c(150, 6.5),
alpha = c(-.95, 0.48),
sigma2u =
  matrix(
    c(93, 12, 0,
      12, 3.8, 0,
      0, 0, 0.4),
    nrow = 3,
    ncol = 3,
    byrow = TRUE
  ),
sigma2e = NULL,
log_sigma2e = TRUE,
linear_sigma2e_attempts = 10,
random_seed = 1
)

simu_no_u2j <- within_ind_var_simu_A(
  how_many_L2_per_L3 = 1000,
  how_many_L1_per_L2 = 9,
  x1_range = c(-1, 1),
  var_names = c("cons", "age"),
  x_matrix_A_x0_x1 = c(TRUE, TRUE),
  z_matrix_A_x0_x1 = c(TRUE, TRUE),
  x_matrix_B_x0_x1 = c(TRUE, TRUE),
  z_matrix_B_x0_x1 = c(FALSE, FALSE),
  beta = c(150, 6.5),
  alpha = c(-.95, 0.48),
  sigma2u =
    matrix(
      c(93, 12,
        12, 3.8),
      nrow = 2,
      ncol = 2,
      byrow = TRUE
    ),
  log_sigma2eij = TRUE,
  linear_sigma2eij_attempts = 10,
  random_seed = 1
)

## Note age uncentred
simu_u2j_covary_linearL1varfun_Revised <- windsimu(
  how_many_L2_per_L3 = 1000,
  how_many_L1_per_L2 = 9,
  time_var_range = c(11.25, 13.25),
  var_names = c("cons", "age"),
  x_matrix_A_cons_time = c(TRUE, TRUE),
  z_matrix_A_cons_time = c(TRUE, TRUE),
  x_matrix_B_cons_time = c(TRUE, TRUE),
  z_matrix_B_cons_time = c(TRUE, FALSE),
  beta = c(150, 6.5),
  alpha = c(-.95, 0.48),
  sigma2u =
    matrix(
      c(93, 12, 2.5,
        12, 3.8, 0.8,
        2.5, 0.8, 0.4),
      nrow = 3,
      ncol = 3,
      byrow = TRUE
    ),
  sigma2e = NULL,
  log_sigma2e = FALSE,
  linear_sigma2e_attempts = 100,
  random_seed = 1
)
}
}
